Stamp of approval (beta) on AgnLocusIndex class
I integrated the LocusPocus program, primarily to test the newly implemented AgnLocusIndex class. This helped identify and fix a few bugs. Testing on a couple of simple data sets suggests that the class is working correctly. I ran valgrind on LocusPocus and fixed a few memory leaks. A couple remain, but they appear to be related to OpenMP and the core GenomeTools code. At any rate, they appear to be unconsequential leaks.