More changes to comparative evaluation code
This commit introduces additional changes to the comparative evaluation code. Many of the data structures have been renamed, and AgnCliquePair class' and dependency on PeComparEval was removed. The AgnPairwiseCompareLocus still depends on PeComparEval, but addressing this dependency will require more involved changes--which of course is the purpose of this branch.