Closes #64.
It turns out that adjusting the number of allowed open file descriptors solved this issue. I've added a check for now to give a helpful warning message. However, I may want to consider whether it is necessary to maintain open filehandles for each sequence. Is the relevant data still in memory near the end of program execution? If so, I can just wait to write the data until then. If not I need to see how much I/O overhead there is from opening, closing, and reopening files.