Fixed an issue with memory management
The AgnLocusStream keeps a pointer to the previous locus (for purposes of computing delta extensions).
This has never been a problem before, since the node stream always keeps a cache (or something like that) anyway, and the AgnLocusStream would never refer to a locus more than 1 step back in the sequence.
However, this changed once we implemented the intron gene handling: the intron gene would never be stored as the previous locus, so the next locus would refer to the outer gene, which by then had been deleted by the node stream.

This commit simply adds a reference count each time a node is stored as the previous node, and subtracts a reference count when it is replaced.